gcf_gcs_to_gbq/requirements.txt
functions_framework==3.4.0
protobuf==4.24.1
google-cloud-bigquery==3.11.4
google-cloud-logging==3.6.0